backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.1.1
google-generativeai==0.8.3
